diarization/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
braceexpand==0.1.7
editdistance==0.6.0
fastcluster==1.2.6
flasgger==0.9.7.1
Flask==2.3.2
h5py==3.7.0
hydra-core==1.2.0
inflect==5.6.0
intervaltree==3.1.0
ipython==8.16.1
ipykernel==6.15.0
jsonlines==3.1.0
kaldi-io==0.9.4
nemo-toolkit==1.17.0
numexpr==2.8.3
numpy==1.24.4
omegaconf==2.2.1
onnx==1.14.1
onnxruntime-gpu==1.16.0
pandas==1.4.2
pyannote.core==4.4
pyannote.database==5.0.1
pyannote.metrics==3.2
pytest==7.3.1
pytorch-lightning==1.8.6
sentencepiece==0.1.95
scikit-learn==1.1.2
scipy==1.9.1
soundfile==0.12.1
tabulate==0.8.10
torch-stft==0.1.4
torch==1.13.1+cu116
torchmetrics==1.2.0
transformers==4.28.1
webdataset==0.1.62
youtokentome==1.0.6
